requirements.txt
agno
openai
python-dotenv
duckduckgo-search
yfinance
youtube-transcript-api
fastapi
uvicorn
python-multipart
